Limit initial values to prevent incompatibilities
JAGS throws an error when the binomial success probability is either 0 or 1 and
the data are not 0/n or n/n. This is not unlikely with the cloglog link, since
a hazard of ~ 3.63 will translate to a probability of 1 due to rounding. It is
much less likely to happen for binomial / cloglog on the low end or for
binomoial / logit, but bounds have been set for those as well.