Change baseline priors approach
The old approach assumed a scale with inbounded support. However, the
log risk ratio is bounded from above at zero, and the rate parameter of
the poisson is bounded from below at zero. The prior is now set on the
event probability for the binomial likelihoods, and directly on the rate
parameter for the Poisson.

Starting values for the baseline effect are (temporarily?) disabled, and
thus automatically generated by JAGS/BUGS. A regression test for
binom/cloglog is failing, possibly because of this.