Fix a bug in `decompose.network()`
`decompose.network()` failes to run when all studies in `network` are multi-arm
trials. That's becauses `filter.network()` call in `decompose.network()` can not
successfully construct the non-existing 2-arm trials network in this scenario.

This patch adds a check for this scenario and directly construct the network
when all studies are multi-arm and filter for 2-arm trials is not necessary.